+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-### 0.2.0
+### 0.2.0 (not released)
 
-* Locked helm version to 2.10.0
-* Added `HELM_MATCH_SERVER` environment variable (see README.md)
+* Locked `helm` version to `2.10.0`
+* Locked `vault` version to `0.10.4`
+* Added `HELM_MATCH_SERVER` environment variable (see [README.md](README.md))